seperated tooltipHandlers to be their own function
