+++ client/src/utils/searchIt.js
@@ -1,6 +1,7 @@
 import searchTicker from "./searchTicker";
 import indicatorIt from "./indicatorIt";
 import { fetchParser } from "./fetchParser";
+import pointRadiiHandler from "./tooltiplabelHandler";
 
 //timeInterval is fed into our API request
 //Handle button click
@@ -15,28 +16,6 @@ const searchIt = async (
   isNews,
   chartData
 ) => {
-  let smallRadiiSize = 1;
-
-  const pointRadiiHandler = (isNews) => {
-    //Check to see if isNews's display is on, and whether there is any information to be fetched.
-    if (isNews?.isDisplayNews && isNews?.currNews.length > 0) {
-      return getNewsRadii(isNews.currNews);
-    } else {
-      return smallRadiiSize;
-    }
-  };
-
-  const getNewsRadii = (newsArr) => {
-    const radii = [];
-    for (let i = 0; i < newsArr.length; i++) {
-      if (newsArr[i] === undefined) {
-        radii.push(smallRadiiSize);
-      } else {
-        radii.push(5);
-      }
-    }
-    return radii;
-  };
   try {
     // console.log("Looking for undefined: ", search, timeInterval, setChartData);
     const response = await searchTicker(search, timeInterval);
